(#30) Refactor formula 3.30
