+++ tests/codes/eurocode/nen_en_1992_1_1_c2_2011/chapter_3_materials/test_formula_3_30.py
@@ -2,7 +2,7 @@
 # pylint: disable=arguments-differ
 import pytest
 
-from blueprints.codes.eurocode.nen_en_1992_1_1_c2_2011.chapter_3_materials import Form3Dot30RatioLossOfPreStressClass3
+from blueprints.codes.eurocode.nen_en_1992_1_1_c2_2011.chapter_3_materials.formula_3_30 import Form3Dot30RatioLossOfPreStressClass3
 
 
 class TestForm3Dot30RatioLossOfPreStressClass3:
